Make implementation match feature request.
Instead of capping the loan rate at the user specified min we do not place
loans less than the min rate at all. This allows the user to transfer and trade
funds to other currencies on long periods of low rates. Without this the bot
would keep recreating the loan at the same rate if cancelled by the user.